modules/group-iv-adaptive-systems/learning-adaptation/pyproject.toml
[project]
name = "learning-adaptation"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "stable-baselines3>=2.3.0",
    "gymnasium>=0.29.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "structlog>=24.1",
    "httpx>=0.27",
    "endogenai-vector-store",
    "endogenai-a2a>=0.1.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "mypy>=1.10.0", "ruff>=0.4.0"]

[tool.uv.sources]
endogenai-vector-store = { path = "../../../shared/vector-store/python", editable = true }
endogenai-a2a = { path = "../../../shared/a2a/python", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/learning_adaptation"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "integration: marks tests requiring live services",
]

[tool.mypy]
strict = true
python_version = "3.11"
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
